+++ src/components/Main/Schedule/LoadingScheduleList/LoadingScheduleList.js
@@ -5,7 +5,7 @@ import LoadingScheduleListEntry from "./LoadingScheduleListEntry";
 const Wrapper = styled.div`
   width: 100%;
   height: 180px;
-  background-color: #eeeeee;
+  background-color: ${({ theme }) => theme.skeletonColor.background};
   display: flex;
   justify-content: space-evenly;
   align-items: center;
